modified: #5 네비게이션 2 -> 3번째 페이지 바로 이동으로 수정
